requirements.txt
Flask
google-generativeai
python-dotenv
Flask-Login
firebase-admin